FROMLIST: nl80211/mac80211: Extend NoAck policy command with peer MAC address
Provides peer level NoAck policy configuration by extending
NL80211_CMD_SET_NOACK_MAP command with peer MAC address.
If user space does not give any peer mac address, the driver
should retain the existing functionality of applying the NoAck
policy for all the staions connected to the netdev. For a connected
peer, the peer specific configuration takes precedence over netdev
level configuration when both are set by the user. When the command
is given without attribute NL80211_ATTR_NOACK_MAP for a peer, the
NoAck policy for the peer will be reset to default which is peer's
netdev NoAck policy. Newly connected stations will have default
NoAck policy. Drivers supporting per-sta NoAck policy must advertise
the support through the extended flag index
NL80211_EXT_FEATURE_PER_STA_NOACK_MAP. Since a value of -1 NoAck map
is used for default configuration, redefine noack_map as int from u16.

(Applied from https://patchwork.kernel.org/patch/10644143/)

Singed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1402056
Tested-by: Julan Hsu <julanhsu@google.com>

+++ include/net/mac80211.h
@@ -3751,7 +3751,7 @@ struct ieee80211_ops {
 			    u8 instance_id);
 
 	int (*set_noack_tid_bitmap)(struct ieee80211_hw *hw,
-				    struct ieee80211_vif *vif, u16 noack_map);
+				    struct ieee80211_vif *vif, int noack_map);
 };
 
 /**
